move performance test to textdb-perftest
