+++ textdb/textdb-perftest/textdb-perftest/pom.xml
@@ -7,8 +7,8 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
     
-    <artifactId>textdb-dataflow</artifactId>
-    <name>textdb-dataflow</name>
+    <artifactId>textdb-perftest</artifactId>
+    <name>textdb-perftest</name>
     <url>http://maven.apache.org</url>
 
     <properties>
@@ -21,6 +21,28 @@
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
             <version>${junit.version}</version>
+        </dependency>
+        
+        <dependency>
+            <groupId>org.apache.lucene</groupId>
+            <artifactId>lucene-core</artifactId>
+            <version>5.5.0</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.lucene</groupId>
+            <artifactId>lucene-analyzers-common</artifactId>
+            <version>5.5.0</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.lucene</groupId>
+            <artifactId>lucene-queryparser</artifactId>
+            <version>5.5.0</version>
+        </dependency>
+        
+        <dependency>
+            <groupId>org.json</groupId>
+            <artifactId>json</artifactId>
+            <version>20160212</version>
         </dependency>
         
         <dependency>
