--- textdb/textdb-perftest/textdb-perftest/src/main/java/edu/uci/ics/textdb/textdb_perftest/App.java
@@ -1,13 +1,0 @@
-package edu.uci.ics.textdb.textdb_perftest;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
